Feat: [BADA-300] 거래 useSearchParams 사용하는 페이지 CSR 분기 구조 개선 및 Suspense 적용 #195
