+++ src/features/trade/all/_client/AllPageClient.tsx
@@ -5,13 +5,10 @@ import { TradeDeadlineBanner } from '@/features/trade/ui/TradeDeadlineBanner';
 import { TradeFlatTab } from '@/features/trade/ui/TradeFlatTab';
 import { TradeList } from '@/features/trade/ui/TradeList';
 import { TradeTrendingBanner } from '@/features/trade/ui/TradeTrendingBanner';
-import { BaseLayout } from '@/shared/ui/BaseLayout';
-import { Header } from '@/shared/ui/Header';
-import { TradeFloatingButton } from '@/widgets/trade/floating-button/ui/TradeFloatingButton';
 import { TradeSearchInput } from '@/widgets/trade/search-input/ui/TradeSearchInput';
 import { TradeSortFilter } from '@/widgets/trade/trade-sort-filter';
 
-export default function AllPage() {
+export default function AllPageClient() {
   const {
     posts,
     isLoading,
@@ -21,20 +18,17 @@ export default function AllPage() {
 
   return (
     <>
-      <BaseLayout header={<Header />} showBottomNav fab={<TradeFloatingButton />} paddingX>
-        <TradeFlatTab basePath="/trade" />
-        <TradeSearchInput />
-        <TradeDeadlineBanner />
-        <TradeTrendingBanner />
-        <TradeList
-          items={posts}
-          isLoading={isLoading}
-          sortLabel={sortOption === 'latest' ? '최신순' : '인기순'}
-          onSortClick={openSortDrawer}
-          onItemClick={onItemClick}
-        />
-      </BaseLayout>
-
+      <TradeFlatTab basePath="/trade" />
+      <TradeSearchInput />
+      <TradeDeadlineBanner />
+      <TradeTrendingBanner />
+      <TradeList
+        items={posts}
+        isLoading={isLoading}
+        sortLabel={sortOption === 'latest' ? '최신순' : '인기순'}
+        onSortClick={openSortDrawer}
+        onItemClick={onItemClick}
+      />
       <TradeSortFilter
         isOpen={isSortDrawerOpen}
         onClose={closeSortDrawer}
