+++ src/features/trade/page/TradePage.tsx
@@ -1,16 +1,11 @@
-'use client';
+import { Suspense } from 'react';
 
-import { useSearchParams } from 'next/navigation';
+import TradePageClient from '@/features/trade/_client/TradePageClient';
 
-import DataPage from '@/features/trade/data/page/DataPage';
-import GifticonPage from '@/features/trade/gifticon/page/GifticonPage';
-import AllPage from '@/features/trade/page/AllPage';
-
-export default function TradePage() {
-  const searchParams = useSearchParams();
-  const currentTab = searchParams?.get('page') ?? 'all';
-
-  if (currentTab === 'data') return <DataPage />;
-  if (currentTab === 'gifticon') return <GifticonPage />;
-  return <AllPage />;
+export default function Page() {
+  return (
+    <Suspense fallback={<div>게시물을 불러오는 중입니다...</div>}>
+      <TradePageClient />
+    </Suspense>
+  );
 }
